Great progress on the kaskawulsh fitting project. Building out script to do all the fitting and plotting for the various accumulation models and fitting approaches
